cleanup on styles and components, removing comments, unused imports
