+++ mdx/posts/example.mdx
@@ -24,6 +24,11 @@ Generated content that I could manage through a [headless CMS](https://www.story
   priority
 />
 
+<Note>
+  This is a note message, dont let it jump into a newline or this will 
+revert to the basic paragraph node
+</Note>
+
 - Oranges
 - Apples
