+++ src/components/Note.tsx
@@ -11,7 +11,7 @@ export const Note = ({
       sx={{ backgroundColor: darken("muted", 0.1) }}
       {...props}
     >
-      {children}
+      <Paragraph>{children}</Paragraph>
     </Message>
   );
 };
