+++ src/pages/_document.tsx
@@ -6,39 +6,8 @@ import Document, {
   DocumentContext,
 } from "next/document";
 import { InitializeColorMode } from "theme-ui";
-import { ServerStyleSheet } from "styled-components";
 
 export default class AppDocument extends Document {
-  // static async getInitialProps(ctx: DocumentContext): Promise<{
-  //   styles: JSX.Element;
-  //   html: string;
-  //   head?: JSX.Element[];
-  // }> {
-  //   const sheet = new ServerStyleSheet();
-  //   const originalRenderPage = ctx.renderPage;
-
-  //   try {
-  //     ctx.renderPage = () =>
-  //       originalRenderPage({
-  //         enhanceApp: (App) => (props) =>
-  //           sheet.collectStyles(<App {...props} />),
-  //       });
-
-  //     const initialProps = await Document.getInitialProps(ctx);
-  //     return {
-  //       ...initialProps,
-  //       styles: (
-  //         <>
-  //           {initialProps.styles}
-  //           {sheet.getStyleElement()}
-  //         </>
-  //       ),
-  //     };
-  //   } finally {
-  //     sheet.seal();
-  //   }
-  // }
-
   static async getInitialProps(ctx: DocumentContext) {
     const initialProps = await Document.getInitialProps(ctx);
     return { ...initialProps };
