+++ src/styles/buttons.ts
@@ -121,46 +121,4 @@ export const buttons = {
       color: "primary",
     },
   },
-  scrollToTop: {
-    backgroundColor: "transparent",
-    padding: 0,
-    color: "grayness",
-    width: "fit-content",
-    cursor: "pointer",
-    "&:hover": {
-      color: "primary",
-    },
-    "&:focus": {
-      color: "primary",
-    },
-  },
-  link: {
-    color: "text",
-    width: "fit-content",
-    boxShadow: "md",
-    marginTop: "auto",
-    cursor: "pointer",
-    textDecoration: "none",
-    backgroundColor: "highlight",
-    "&:hover": {
-      backgroundColor: "primary",
-      color: "black",
-    },
-    "&:focus": {
-      outline: "none",
-      transition: "box-shadow 0.2s",
-      boxShadow: "0px 0px 1px 3px #4299e1",
-    },
-  },
-  navButton: {
-    cursor: "pointer",
-    textDecoration: "none",
-    color: "text",
-    "&:hover": {
-      color: "primary",
-    },
-    "&:focus": {
-      color: "primary",
-    },
-  },
 };
