--- src/styles/links.ts
@@ -1,10 +1,0 @@
-export const links = {
-  card: {
-    zIndex: 2,
-    fontFamily: "body",
-    fontSize: [2, 3],
-    width: "fit-content",
-    textDecoration: "none",
-    "&:hover": { textDecoration: "underline" },
-  },
-};
