--- src/styles/text.ts
@@ -1,30 +1,0 @@
-export const text = {
-  default: {
-    color: "text",
-    fontFamily: "body",
-    fontWeight: "body",
-    lineHeight: "body",
-    fontSize: [3, 4],
-  },
-  paragraph: {
-    color: "text",
-    fontFamily: "body",
-    fontWeight: "body",
-    lineHeight: "body",
-    fontSize: [3, 4],
-  },
-  heading: {
-    color: "text",
-    fontFamily: "heading",
-    fontWeight: "heading",
-    lineHeight: "heading",
-  },
-  main: {
-    fontSize: [4, 5],
-    fontWeight: "bold",
-  },
-  subheader: {
-    fontSize: [2, 3],
-    fontWeight: "body",
-  },
-};
